requirements.txt
numpy>=1.24.0
opencv-python>=4.8.0
scikit-learn>=1.3.0
scikit-image>=0.22.0
jupyter>=1.0.0
matplotlib>=3.7.0
pillow>=10.0.0
torch>=2.1.0
torchvision>=0.16.0
torchmetrics>=1.3.0
torch-fidelity>=0.3.0
lpips>=0.1.4
streamlit>=1.35.0
